Add windows warning fixes

+++ source/targets/salesman_angle/source/Position.cpp
@@ -23,8 +23,8 @@ IMove::CPtr SalesmanAnglePosition::generateMove() const
   if (cities.size() < 2) {
     return std::make_unique<SalesmanAngleMove>(0, 0, 0);
   }
-  std::size_t idx1 = r.randomInt(0, cities.size() - 1);
-  std::size_t idx2 = r.randomInt(0, cities.size() - 2);
+  std::size_t idx1 = r.randomInt(0, int(cities.size()) - 1);
+  std::size_t idx2 = r.randomInt(0, int(cities.size()) - 2);
   if (idx2 >= idx1) {
     ++idx2;
   } else {
@@ -68,8 +68,8 @@ void SalesmanAnglePosition::makeMove(IMove::CPtr move)
 
 int SalesmanAnglePosition::size() const
 {
-  return sizeof(double) + sizeof(std::vector<std::pair<double, double>>) +
-         sizeof(std::pair<double, double>) * cities.capacity();
+  return int(sizeof(double) + sizeof(std::vector<std::pair<double, double>>) +
+         sizeof(std::pair<double, double>) * cities.capacity());
 }
 
 IPosition::CPtr SalesmanAnglePosition::clone() const { return std::make_unique<SalesmanAnglePosition>(energy, cities); }
